Refactored RedisScoreLimit to ScoreLimitOptions.
