+++ java/integTest/src/test/java/glide/SharedCommandTests.java
@@ -20,8 +20,8 @@ import glide.api.BaseClient;
 import glide.api.RedisClient;
 import glide.api.RedisClusterClient;
 import glide.api.models.commands.ExpireOptions;
-import glide.api.models.commands.RedisScoreLimit.InfBound;
-import glide.api.models.commands.RedisScoreLimit.ScoreBoundary;
+import glide.api.models.commands.ScoreLimitOptions.InfBound;
+import glide.api.models.commands.ScoreLimitOptions.ScoreBoundary;
 import glide.api.models.commands.SetOptions;
 import glide.api.models.commands.ZaddOptions;
 import glide.api.models.configuration.NodeAddress;
